Move handler into utils, dont trigger for inputs
