+++ src/data-renderers/__tests__/Table.test.js
@@ -122,57 +122,5 @@ describe('Table', () => {
       instance.handleSelectAll()
       expect(props.select).toHaveBeenCalledWith([2, 4, 5])
     })
-    it('handles clicking on row', () => {
-      spyOn(props, 'itemOnClick')
-      const row = {}
-      const currentTarget = {}
-      const target = {
-        parentElement: currentTarget,
-        tagName: 'TD',
-      }
-      const mockEvent = {
-        currentTarget,
-        target,
-        persist: jasmine.createSpy(),
-        preventDefault: jasmine.createSpy(),
-        stopPropagation: jasmine.createSpy(),
-      }
-      const instance = shallow(
-        <Table
-          {...props}
-        />
-      ).instance()
-      instance.handleRowClick(mockEvent, row)
-      expect(mockEvent.persist).toHaveBeenCalled()
-      expect(mockEvent.preventDefault).toHaveBeenCalled()
-      expect(mockEvent.stopPropagation).toHaveBeenCalled()
-      expect(props.itemOnClick).toHaveBeenCalledWith(row)
-    })
-    it('handles clicking on a link in a row', () => {
-      spyOn(props, 'itemOnClick')
-      const row = {}
-      const currentTarget = {}
-      const target = {
-        parentElement: currentTarget,
-        tagName: 'A',
-      }
-      const mockEvent = {
-        currentTarget,
-        target,
-        persist: jasmine.createSpy(),
-        preventDefault: jasmine.createSpy(),
-        stopPropagation: jasmine.createSpy(),
-      }
-      const instance = shallow(
-        <Table
-          {...props}
-        />
-      ).instance()
-      instance.handleRowClick(mockEvent, row)
-      expect(mockEvent.persist).toHaveBeenCalled()
-      expect(mockEvent.preventDefault).not.toHaveBeenCalled()
-      expect(mockEvent.stopPropagation).not.toHaveBeenCalled()
-      expect(props.itemOnClick).not.toHaveBeenCalled()
-    })
   })
 })
