+++ src/data-renderers/utils/index.js
@@ -1,3 +1,4 @@
 export {
   getValueFromAccessor,
+  handleRowClick,
 } from './functions'
